requirements.txt
flask
flask-cors
pymongo
python-dotenv
certifi
gunicorn